requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
asarPy==1.0.1
asyncpg==0.30.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distlib==0.3.9
docker==7.1.0
fastapi==0.115.5
filelock==3.16.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
Mako==1.3.6
MarkupSafe==3.0.2
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pydantic==2.10.1
pydantic_core==2.27.1
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
virtualenv==20.27.1
